[vm] Replace VmRegions with VmAddressRegions
VmAddressRegions are the new VM partitioning layer.  They are
a tree-structured partitioning of the address space in which
each region can be subdivided into additional regions and VM
mappings.  This gives the caller more control over the address
space layout, such as being able to guarantee that parts of the
address space won't be used by anything else.

This patch implements the core of VMARs, but does not expose them
yet.  It includes a compatibility layer in VmAspace for users of
the old interface.

This drops support for min_alloc_gap, which will be unnecessary
once the migration is finished.

This changes the behavior of VMM_FLAG_VALLOC_BASE, to actually
just be a hint to start the search for space near the top of the
user aspace.

The following should be coming in followup patches:
- Restricting invocation of the compatibility layer to the kernel
  address space only
- Exposing new syscalls to allow usermode to manipulate VMARs
- Integrating those syscalls into our process creation flow
- Replacing the allocation scheme within the VMAR code to be randomized
  rather than first-fit placement.

Change-Id: I550a3120320d4f6c4cf72caeb4ed3226c79104f7